+++ spec/features/staff/users_can_create_a_transaction_spec.rb
@@ -60,11 +60,24 @@ RSpec.feature "Users can create a transaction" do
       expect(page).to have_content("Transaction type can't be blank")
       expect(page).to have_content("Date can't be blank")
       expect(page).to have_content("Value must be between 1 and 99,999,999,999.00")
-      expect(page).to have_content("Disbursement channel can't be blank")
       expect(page).to have_content("Providing organisation name can't be blank")
       expect(page).to have_content("Providing organisation type can't be blank")
       expect(page).to have_content("Receiving organisation name can't be blank")
       expect(page).to have_content("Receiving organisation type can't be blank")
+    end
+
+    scenario "disbursement channel is optional" do
+      activity = create(:fund_activity, organisation: user.organisation)
+
+      visit organisation_path(user.organisation)
+
+      click_on(activity.title)
+
+      click_on(I18n.t("page_content.transactions.button.create"))
+      expect(page).to have_content("Disbursement channel (optional)")
+      click_on(I18n.t("generic.button.submit"))
+
+      expect(page).to_not have_content("Disbursement channel can't be blank")
     end
 
     context "Value number validation" do